event

+++ src/event_flags.rs
@@ -1,5 +1,5 @@
 bitflags! {
-        flags EventFlags: u32 {
+        flags EventFlags: u64 {
             const FLAG_TIMEOUT 	 = 0b00000001,
             const FLAG_READ      = 0b00000010,
             const FLAG_WRITE     = 0b00000100,
